[8Days] Rebuild svg-sprite
As the main logo and the phone icon have a linear-grad it has trouble with displaying in FF and IE if used with svg+use. Could not find a basic solurion so I decided to rebuid the sprite and replace svg+use with simple img tags. This is where Pug comes in handy, this procedure would take me ages to fix by hand on all pages. What a drag :low_brightness: